Merge README from bootable application
This commit rewords and merges the instructions removed from the corresponding bootable application. It also introduces new instructions to cover the new demo update mechanism: flashing the update binary to a reserved area in internal flash to simulate an update.

This commit also temporarily retargets the mcuboot.lib dependency to a branch of mcuboot that includes yet-to-be merged commits.